Defer converting Rack env variables to RubyStrings until needed
To accomplish this we had to subclass RubyHash to RackEnvironmentHash
and override all methods that are exposed via Ruby. It may become a
burden to maintain in the future but it prevents us from taking the
cost of converting all Rack environment keys from Java to Ruby for
every request.