Fix the RubyLanguage runtime logic to only use the global runtime once
After the RubyLanguage is used from a container once, we have to use
new JRuby runtimes for any future usage. In the embedded case, we'll
only be creating one JRuby runtime so we'll always be using the global
one. But, for testing, we create and destroy many container instances
and each one needs a new Ruby runtime for isolation.