Allow modules to specify the languages/components they provide.
Modules now ship a META-INF/wunderboss.properties of the form:

    language.bumpkin=org.bumpkin.BumpkinLanguage
    component.biscuit=com.biscuitude.BiscuitComponent
    ...

The default WunderBoss instance will then load all of the
META-INF/wunderboss.properties files on the effective classpath and
search for the proper language or component class if it does not
already have a registered language/component of the requested type.

WunderBoss#hasLanguage/hasComponent will now also trigger this load
and search, to allow dependencies to be automatically resolved.