Allow wunderboss-wildfly.jar to be included outside of WildFly
This changes does a couple of things:

1) No longer shades any artifacts into wunderboss-wildfly.jar

2) Mark all wunderboss dependencies of wunderboss-wildfly as provided
and catch LinkageErrors when trying to reference things that depend on
the web and messaging modules. Long-term, we need to probably move
WildFlyWeb and WildFlyMessaging out of the wildfly module and into
the web/messaging modules.

3) Removes wunderboss-wildfly's wunderboss.properties in favor of
manually registering all the WildFly-specific components in
WildFlyService. This allows wunderboss-wildfly.jar to be on the
classpath without forcing WildFly versions of components to be used,
like when running from an executable jar.

3) Look for a META-INF/app.properties if we can't find an external
properties file named the same as the deployable jar. This is to
support fully self-contained jar artifacts.