Expose USERNAME, PASSWORD, and REMOTE_TYPE messaging connection options.
If the remote HornetQ is in WildFly, it requires clients to use http
upgrade. There is currently no good way to detect if the remote HQ
requires http upgrade, and specifying it on the connection factory isn't
compatible with a non-http-upgrade server. To handle this, we require
passing the REMOTE_TYPE as "hornetq-wildfly" when connecting to an in-WF
HQ.

For connecting to a standalone, non-http-upgrade HQ, REMOTE_TYPE is not
required.