Use slf4j and logback for logging.
Since quartz requires slf4j, we'll use it as our logging abstraction
everywhere instead of jboss-logging. And, since we want logging to work
out-of-the-box, we need a logging implementation. We were bringing in
log4j, but have switched to logback since it requires less dependencies
to work as an slf4j implementation, and it is the successor to log4j.

Known issues:

* If a user wants to use a different logging provider, they'll need to
  exclude logback and bring in their own. If they don't exclude, it may
  still work, but they'll get a stern message from slf4j about multiple
  implementations.

* The messaging tests now throw "java.lang.ClassNotFoundException:
  org.jboss.logmanager.LogManager", but this doesn't affect the tests,
  nor does it show up for consumers of wunderboss-messaging. I don't
  currently know the cause of this message.