Don't touch HornetQ directly in the container.
This uses MSC services to create/destroy destinations instead of using
the JMSServerManager. This allows us to keep the HornetQ module out of
the application, allowing things that bring in their own conflicting
HornetQ (like Datomic) to work.

This also adds the ability for destinations to be looked up by jndi name
passed as the name to lookupOrCreate*.