Collapse connection and session to context.
This embraces the JMS 2.0 concept of a context, which takes the place of
a connection and session pair. We now expose a Context instead of a
Connection and Session, and now no longer have the concept of default
connection. If a user needs the performance gain of connection/session
sharing, they need to self-manage a context.

This also renames all of the HornetQ* classes we provide to HQ* to avoid
annoying conflicts with HornetQ's own classes, and renames send() to
publish() to match what TorqueBox and Immutant expose.