Fixes to allow an external JRuby to work for a jar.
The important change here is setting the ClassLoader on the JRuby
instance to the WunderBoss loader - that allows JRuby to see things
loaded from an external jar.

This also changes the jruby-home lookup order, allowing an external
JRuby to override an included one, and throws a helpful exception if
JRuby can't be loaded.