Provide an arjuna-properties.xml to fix [IMMUTANT-548]
Narayana relies on a config file in its classpath, the name of which is
configured in a manifest entry, because you can't have too many levels
of indirection. In the narayana-jta jar, this file is named
jbossts-properties.xml, but when that jar is packaged into an uberjar,
the manifest attribute is lost.

Fortunately, Narayana's ConfigurationInfo class will fall back to
looking for a resource named arjuna-properties.xml, so we've copied the
jbossts-properties.xml file packaged in the narayana-jta jar into the
wunderboss-transactions jar under the name, arjuna-properties.xml

Both will then be found in the uberjar but, lacking the manifest
attribute, only the arjuna-properties.xml will be used.