Expect the delegate endpoint to be stored under the thread id [IMMUTANT-554]
We can't use a static key for the endpoint in the config's
userProperties, because that map is shared for all requests, which leads
to getting the wrong endpoint under the slightest load.

We can use the thread id here because onOpen will be triggered on the
same thread as modifyHandshake, where we store the endpoint.