Fix classpath splitting on Windows
We delimit classpaths in the properties file inside our jar files with
a colon. But, on Windows, a colon is also used to separate drive
letters from the rest of the path. So, we detect the case where we
split a path up into its drive letter and path and reassemble that
into the correct path.

We can't just rely on File.pathSeparator because then the built Jar
would only work on the same OS Family (Windows vs Linux/Mac) that it
was created on.