Ensure onClose gets called when underlying channel is closed [IMMUTANT-563, IMMUTANT-564]
If the client doesn't send a close frame, we weren't calling onClose
when outside of WildFly. This registers a listener on the channel that
gets called on close that then ensures onClose gets called. This also
fixes the issue where onClose wasn't getting called if the server
instance was stopped server-side.