Introduce module to make caching compatible with WF8
Because we upgraded Infinispan to 7 to fix IMMUTANT-454, we broke
WildFly 8, which ships with Infinispan 6, because the Cache interface
changed.

So now we have a new module that addresses the incompatibilities by
making CacheWithCodec abstract and introducing an Encoder interface,
which works with 7 by default (out of container and in WF9). But the
wildfly module now detects WF8 to trigger the use of an Encoder
compatible with the Cache interface in ispan6.