First pass at caching support in EAP 6.4 [IMMUTANT-569]
We introduce a bit of reflective hackiness in Config to account for
ispan 5 not having keyEquivalence and loaders instead of persistence
builders.

The lack of keyEquivalence means byte[]'s can't be used as keys, so we
override withCodec in WildFlyCaching to wrap byte[] keys with smarter
hashCode/equals methods.

We may convert this to a Codec in the future, but doing so comes at the
cost of needlessly encoding values instead of keys. At least, as the
code is currently structured.