Don't allow setting a stream timeout on WF 8.2
Setting a stream timeout triggers the same session deadlock bug as async
sending in WF 8.2, so we throw if you try to use it in 8.2. This
requires changing the init order of the stream, because if we throw
after the request is converted to async, we can never return a response
to the client.