Break wildfly up in to multiple modules
This is the first step in allowing wunderboss-based apps to properly
take advantage of swarm, since wildfly functionality can be brought in
piecemeal, instead of as one dep. This also means that we don't assume a
full WF is available, and each piece only assumes what it needs is
there.

There is still an org.projectodd.wunderboss/wunderboss-wildfly, but it
is an aggregator for all of the other WF modules.