Use the simple election policy for singletons when in WF 8/9
There is a bug (https://issues.jboss.org/browse/WFLY-5108) that can
cause non-deterministic election policies to result in an inconsistent
singleton state (multiple nodes running the singleton, or no nodes
running the singleton are both possible). This caused an intermittent
integ failure in Immutant's cluster tests. We're still using the random
election policy with WF10 - the Immutant tests should catch it if it
also has the same behavior.