Auto-close async HEAD requests after the first send [IMMUTANT-630]
If a HEAD request is handled by something that creates an async channel,
the response should use the same headers and status as a GET
would. However, if the user detects the HEAD and returns an empty body,
the response will have a 'Content-Length: 0' header instead of the
'Transfer-Encoding: Chunked' header that a GET would get, which is a
violation of the HTTP
spec (https://tools.ietf.org/html/rfc7230#section-3.3.2). And if we let
the regular channel semantics occur, we may leak connections since they
may not get closed (and there's no value in keeping them open beyond the
first send, since the body of HEAD responses are ignored).